Reorganize some newlib definitions into a "generic" module
